+++ src/photos/pageHelpers/getPage.tsx
@@ -1,42 +1,10 @@
 import Layout from '@src/components/common/Layout'
-import countryGallery from '@src/galleries/uzbekistan'
-import { GalleryIndexPage } from '@src/photos/components/GalleryPage'
 import { SinglePhotoPage } from '@src/photos/components/SinglePhotoPage'
 import { getPhotoIdFromRouter } from '@src/photos/pageHelpers/getPhotoIdFromRouter'
-import type { CityGallery, CityGalleryMap, CountryGallery, ImageProps, PageProps } from '@src/photos/utils/types'
+import type { CityGalleryMap, CountryGallery, PageProps } from '@src/photos/utils/types'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-
-function GalleryPage({
-    gallery,
-    cityGalleryMap,
-    images,
-}: {
-    gallery: CountryGallery | CityGallery
-    cityGalleryMap: CityGalleryMap
-    images: ImageProps[]
-}) {
-    return (
-        <Layout title={gallery.title} keywords={gallery.keywords}>
-            <h1>{gallery.title}</h1>
-
-            <section>
-                <h2>Photos and Video</h2>
-
-                <p>
-                    <Link href={`/travel/${countryGallery.country}`}>{`Back to ${countryGallery.title}`}</Link>
-                </p>
-
-                <GalleryIndexPage
-                    gallery={gallery}
-                    city={'city' in gallery ? gallery.city : undefined}
-                    cityGalleryMap={cityGalleryMap}
-                    images={images}
-                />
-            </section>
-        </Layout>
-    )
-}
+import { GalleryPage } from '../components/GalleryPage'
 
 export function getPage(countryGallery: CountryGallery, cityGalleryMap: CityGalleryMap) {
     return function Page({ images, currentPhoto }: PageProps) {
